move explore store to separate file
