+++ apps/client/app/(navbar)/explore/store.ts
@@ -1,0 +1,9 @@
+import { create } from "zustand";
+
+export interface IExploreStore {
+  filter: string;
+}
+
+export const useExploreStore = create<IExploreStore>(() => ({
+  filter: "",
+}));
